DEBUG: debugging the profiledatacontewxt.js
