+++ src/contexts/ProfileDataContext.js
@@ -1,17 +1,8 @@
-/* eslint-disable */ 
-// React hooks
-import { React, createContext, useContext, useEffect, useState } from "react";
-
-// Axios
+import { createContext, useContext, useEffect, useState } from "react";
 import { axiosReq, axiosRes } from "../api/axiosDefaults";
-
-// Currentuser context
 import { useCurrentUser } from "../contexts/CurrentUserContext";
-
-// Utils imports
 import { followHelper, unfollowHelper } from "../utils/utils";
 
-// Create the Profile context thats used on the site
 const ProfileDataContext = createContext();
 const SetProfileDataContext = createContext();
 
@@ -19,14 +10,41 @@ export const useProfileData = () => useContext(ProfileDataContext);
 export const useSetProfileData = () => useContext(SetProfileDataContext);
 
 export const ProfileDataProvider = ({ children }) => {
-  const [profileData, setProfileData] = useState({
-    pageProfile: { results: [] },
-    popularProfiles: { results: [] },
-  });
+    const [profileData, setProfileData] = useState({
+        // we will use the pageProfile later
+        pageProfile: { results: [] },
+        popularProfiles: { results: [] },
+      });
+     
+      const currentUser = useCurrentUser();
+     
+      useEffect(() => {
+        const handleMount = async () => {
+          try {
+            const { data } = await axiosReq.get(
+              "/profiles/?ordering=-followers_count"
+            );
+            setProfileData((prevState) => ({
+              ...prevState,
+              popularProfiles: data,
+            }));
+          } catch (err) {
+            console.log(err);
+          }
+        };
+        handleMount();
+      }, [currentUser]);
 
-  const currentUser = useCurrentUser();
 
-  // Handles the clicked profile for following
+
+
+
+
+
+
+
+
+  // 
   const handleFollow = async (clickedProfile) => {
     try {
       const { data } = await axiosRes.post("/followers/", {
@@ -48,14 +66,14 @@ export const ProfileDataProvider = ({ children }) => {
         },
       }));
     } catch (err) {
-      // console.log(err);
+      console.log(err);
     }
   };
 
-  // Handles the clicked profile for unfollowing
   const handleUnfollow = async (clickedProfile) => {
     try {
       await axiosRes.delete(`/followers/${clickedProfile.following_id}/`);
+
       setProfileData((prevState) => ({
         ...prevState,
         pageProfile: {
@@ -71,7 +89,7 @@ export const ProfileDataProvider = ({ children }) => {
         },
       }));
     } catch (err) {
-      // console.log(err);
+      console.log(err);
     }
   };
 
@@ -86,7 +104,7 @@ export const ProfileDataProvider = ({ children }) => {
           popularProfiles: data,
         }));
       } catch (err) {
-        // console.log(err);
+        console.log(err);
       }
     };
