+++ src/pages/profiles/PopularProfiles.js
@@ -1,63 +1,52 @@
-import React, { useEffect, useState } from "react";
-import { Container } from "react-bootstrap";
-import { axiosReq } from "../../api/axiosDefaults";
+/* eslint-disable */
+// React import
+import React from "react";
+
+// Bootstrap
+import Container from "react-bootstrap/Container";
+
+// Styles and CSS
 import appStyles from "../../App.module.css";
+
+// Assets, Profiledata context, also the weather widget library
 import Asset from "../../components/Asset";
-import { useCurrentUser } from "../../contexts/CurrentUserContext";
+import { useProfileData } from "../../contexts/ProfileDataContext";
 import Profile from "./Profile";
 
+
+// Display the most popular profiles and also the weatherwidget
+
 const PopularProfiles = ({ mobile }) => {
-  const [profileData, setProfileData] = useState({
-    // we will use the pageProfile later!
-    pageProfile: { results: [] },
-    popularProfiles: { results: [] },
-  });
-  const { popularProfiles } = profileData;
-  const currentUser = useCurrentUser();
-
-  useEffect(() => {
-    const handleMount = async () => {
-      try {
-        const { data } = await axiosReq.get(
-          "/profiles/?ordering=-followers_count"
-        );
-        setProfileData((prevState) => ({
-          ...prevState,
-          popularProfiles: data,
-        }));
-      } catch (err) {
-        console.log(err);
-      }
-    };
-
-    handleMount();
-  }, [currentUser]);
+  const { popularProfiles } = useProfileData();
 
   return (
-    <Container
-      className={`${appStyles.Content} ${
-        mobile && "d-lg-none text-center mb-3"
-      }`}
-    >
-      {popularProfiles.results.length ? (
-        <>
-          <p>Most followed profiles.</p>
-          {mobile ? (
-            <div className="d-flex justify-content-around">
-              {popularProfiles.results.slice(0, 4).map((profile) => (
-                <Profile key={profile.id} profile={profile} mobile />
-              ))}
-            </div>
-          ) : (
-            popularProfiles.results.map((profile) => (
-              <Profile key={profile.id} profile={profile} />
-            ))
-          )}
-        </>
-      ) : (
-        <Asset spinner />
-      )}
-    </Container>
+    <>
+      <Container
+        className={`${appStyles.Content} ${
+          mobile && "d-lg-none text-center mt-5 mb-3"
+        }`}
+      >
+        
+        {popularProfiles.results.length ? (
+          <>
+            <p>Most followed profiles.</p>
+            {mobile ? (
+              <div className="d-flex justify-content-around">
+                {popularProfiles.results.slice(0, 4).map((profile) => (
+                  <Profile key={profile.id} profile={profile} mobile />
+                ))}
+              </div>
+            ) : (
+              popularProfiles.results.map((profile) => (
+                <Profile key={profile.id} profile={profile} />
+              ))
+            )}
+          </>
+        ) : (
+          <Asset spinner />
+        )}
+      </Container>
+    </>
   );
 };
